requirements.txt
urllib3<2.0
requests
discord.py
python-dotenv
curl-cffi